CLIENT-1024 Support scan/query consistency validation using cluster key.  Enable validation with fail_on_cluster_change policy.

+++ src/main/aerospike/as_tls.c
@@ -405,7 +405,7 @@ password_cb(char* buf, int size, int rwflag, void* udata)
 		return 0;
 	}
 
-	int pw_len = strlen(pw);
+	int pw_len = (int)strlen(pw);
 
 	if (pw_len > size) {
 		return 0;
